chore: rename service layer files
