+++ bc_obps/registration/api/select_operator/request_access.py
@@ -1,4 +1,4 @@
-from service.select_operator.RequestAccessService import RequestAccessService
+from service.select_operator.request_access_service import RequestAccessService
 from registration.decorators import authorize
 from registration.schema import (
     SelectOperatorIn,
