+++ bc_obps/service/select_operator/user_operator/get_user_operator_service.py
