+++ bc_obps/service/user_operator/get_user_operator_admin_exists_service.py
