+++ bc_obps/service/user_operator/get_user_operator_id_service.py
