+++ bc_obps/service/user_operator/get_user_operator_list_from_user_service.py
